requirements-production.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations

# Redis & Caching
redis[hiredis]==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# MCP
mcp[cli]>=1.0.0

# HTTP Client
httpx==0.26.0

# LLM Providers
openai==1.10.0
anthropic==0.18.0

# Monitoring & Observability
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0
httpx==0.26.0  # For testing async HTTP

# Load Testing
locust==2.20.0

# Development
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Rate Limiting
slowapi==0.1.9

# WebSockets
websockets==12.0

# Background Tasks
celery[redis]==5.3.6
flower==2.0.1  # Celery monitoring

# File Processing (for rich media support)
python-magic==0.4.27
pillow==10.2.0
pypdf2==3.0.1

# Data Validation
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Environment & Config
environs==11.0.0

# Database Connection Pooling
psycopg2-binary==2.9.9  # For Alembic CLI

# AWS SDK (optional - for S3, SES, etc.)
# boto3==1.34.27

# Google Cloud (optional - for GCS, etc.)
# google-cloud-storage==2.14.0
